REST API improvments and config file support
The REST API can be started using the -w command line option.
A HTTP GET to this port will return a JSON document with information
about clients, brokers, and subscriptions

The configuration file can be used with the  -f command line option.
The file follows the czmq format and can be used instead of all the
current command line options. An example is included in br0.cfg